requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Google APIs (Slides, Gmail, Drive)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Visualization
plotly>=5.17.0
kaleido>=0.2.1

# Workflow comparison and validation
deepdiff>=6.0.0
jsonschema>=4.20.0

# CLI formatting
rich>=13.0.0
tabulate>=0.9.0

# AI tools
tiktoken>=0.5.0
openai>=1.0.0

# Templating
jinja2>=3.1.0

# Data validation
pydantic>=2.0.0
